[QueryContext] Enhance Predicate.toString() to return valid expression
- The expression returned can be parsed back to the same predicate
- The expression returned is the standardized string representation of the predicate
- This change can help solve the predicate standardization problem in DistinctCountThetaSketch

Also extract the common converter methods in BrokerRequestToQueryContextConverter into QueryContextConverterUtils so that they can be used by other classes
